pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "av-sim-arena"
version = "0.1.0"
description = "Multi-scenario simulation benchmark for autonomous vehicle decision-making"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "AV-Sim-Arena Contributors"},
]
keywords = ["autonomous-vehicles", "simulation", "benchmark", "planning", "safety"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "matplotlib>=3.7",
    "pyyaml>=6.0",
    "fastapi>=0.100",
    "uvicorn>=0.22",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "httpx>=0.24",
    "ruff>=0.1",
]
carla = [
    "carla>=0.9.14",
]
sumo = [
    "traci>=1.18",
]

[project.urls]
Homepage = "https://github.com/av-sim-arena/av-sim-arena"
Repository = "https://github.com/av-sim-arena/av-sim-arena"
Documentation = "https://github.com/av-sim-arena/av-sim-arena#readme"

[tool.setuptools.packages.find]
include = ["av_sim_arena*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py310"
line-length = 100
